# vulkan: Optimize some mat-vec mul quant shaders