@0cc4m 2024-11-14
Thank you, this is quite impressive!

I tested these models:
- Llama 3 8B Q4_K_S
- Llama 3 8B q4_0
- Llama 2 13B q4_0
- Mistral Nemo q5_0

RTX 3090:
```
Before:
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         66.12 ± 0.99 |
| llama 8B Q4_0                  |   5.61 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         57.64 ± 0.04 |
| llama 13B Q4_0                 |   6.86 GiB |    13.02 B | Vulkan     |  99 |         tg128 |         38.73 ± 0.12 |
| llama 13B Q5_0                 |   7.93 GiB |    12.25 B | Vulkan     |  99 |         tg128 |         20.79 ± 0.04 |

After:
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         77.54 ± 0.13 |
| llama 8B Q4_0                  |   5.61 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         73.46 ± 0.09 |
| llama 13B Q4_0                 |   6.86 GiB |    13.02 B | Vulkan     |  99 |         tg128 |         50.04 ± 0.05 |
| llama 13B Q5_0                 |   7.93 GiB |    12.25 B | Vulkan     |  99 |         tg128 |         38.89 ± 0.11 |
```

Tesla P40:
```
Before:
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         21.52 ± 0.03 |
| llama 8B Q4_0                  |   5.61 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         22.80 ± 0.02 |
| llama 13B Q4_0                 |   6.86 GiB |    13.02 B | Vulkan     |  99 |         tg128 |         13.85 ± 0.00 |
| llama 13B Q5_0                 |   7.93 GiB |    12.25 B | Vulkan     |  99 |         tg128 |          8.49 ± 0.01 |

After:
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         38.30 ± 0.12 |
| llama 8B Q4_0                  |   5.61 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         22.19 ± 0.01 |
| llama 13B Q4_0                 |   6.86 GiB |    13.02 B | Vulkan     |  99 |         tg128 |         13.05 ± 0.00 
| llama 13B Q5_0                 |   7.93 GiB |    12.25 B | Vulkan     |  99 |         tg128 |          8.92 ± 0.59 |
```

AMD Radeon Pro VII:
```
Before:
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         25.14 ± 0.52 |
| llama 8B Q4_0                  |   5.61 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         28.82 ± 0.54 |
| llama 13B Q4_0                 |   6.86 GiB |    13.02 B | Vulkan     |  99 |         tg128 |         16.30 ± 1.93 |
| llama 13B Q5_0                 |   7.93 GiB |    12.25 B | Vulkan     |  99 |         tg128 |         14.74 ± 0.06 |

After:
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         29.90 ± 0.03 |
| llama 8B Q4_0                  |   5.61 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         38.49 ± 3.24 |
| llama 13B Q4_0                 |   6.86 GiB |    13.02 B | Vulkan     |  99 |         tg128 |         26.55 ± 0.35 |
| llama 13B Q5_0                 |   7.93 GiB |    12.25 B | Vulkan     |  99 |         tg128 |         18.09 ± 0.21 |
```