@netrunnereve 2024-11-15
This is 50% faster on Q4_0 with a RX 570, very nice! 

**Master**
| model                          |       size |     params | backend    | ngl | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------------: | -------------------: |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         pp512 |         95.83 ± 0.19 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          7.74 ± 0.06 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     | 100 |       8 |         pp512 |         76.86 ± 0.09 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          8.09 ± 0.02 |

**PR**
| model                          |       size |     params | backend    | ngl | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------------: | -------------------: |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         pp512 |         96.81 ± 0.09 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         11.61 ± 0.05 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     | 100 |       8 |         pp512 |         76.83 ± 0.43 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          9.18 ± 0.02 |