@0cc4m 2024-11-15
We also didn't break Intel yet:

Intel ARC A770:
Before:

| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         22.56 ± 0.05 |
| llama 8B Q4_0                  |   5.61 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         28.77 ± 0.05 |
| llama 13B Q4_0                 |   6.86 GiB |    13.02 B | Vulkan     |  99 |         tg128 |         20.51 ± 0.03 |
| llama 13B Q5_0                 |   7.93 GiB |    12.25 B | Vulkan     |  99 |         tg128 |          4.87 ± 0.00 |
| llama 13B Q4_0                 |   6.60 GiB |    12.25 B | Vulkan     |  99 |         tg128 |         21.41 ± 0.08 |

After:

| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         28.89 ± 0.09 |
| llama 8B Q4_0                  |   5.61 GiB |     8.03 B | Vulkan     |  99 |         tg128 |         34.64 ± 0.04 |
| llama 13B Q4_0                 |   6.86 GiB |    13.02 B | Vulkan     |  99 |         tg128 |         21.22 ± 0.01 |
| llama 13B Q5_0                 |   7.93 GiB |    12.25 B | Vulkan     |  99 |         tg128 |          4.74 ± 0.00 |
| llama 13B Q4_0                 |   6.60 GiB |    12.25 B | Vulkan     |  99 |         tg128 |         22.46 ± 0.02 ||

It got a little faster in most cases, but Intel is still quirky. Besides low prompt processing cause ANV doesn't like my mul_mm shader, it seems to have an issue with q5_0 Matrix Vector Multiplication too. I added q4_0 to make sure the issue is not the model. Maybe you have an idea what's going on. Alignment issue with the quant struct? I've had those in the past.

Here's most of the quants on a 1.1B model on Intel A770. Something's definitely wrong with Q5_0 on ARC.

| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
| llama 1B F16                   |   2.05 GiB |     1.10 B | Vulkan     |  99 |         tg128 |         86.78 ± 0.20 |
| llama 1B Q4_0                  | 606.53 MiB |     1.10 B | Vulkan     |  99 |         tg128 |        123.14 ± 0.10 |
| llama 1B Q4_1                  | 668.18 MiB |     1.10 B | Vulkan     |  99 |         tg128 |        113.38 ± 0.18 |
| llama 1B Q5_0                  | 729.84 MiB |     1.10 B | Vulkan     |  99 |         tg128 |         45.57 ± 0.03 |
| llama 1B Q5_1                  | 791.50 MiB |     1.10 B | Vulkan     |  99 |         tg128 |        108.84 ± 0.35 |
| llama 1B Q2_K - Medium         | 411.41 MiB |     1.10 B | Vulkan     |  99 |         tg128 |         98.81 ± 0.28 |
| llama 1B Q3_K - Medium         | 523.67 MiB |     1.10 B | Vulkan     |  99 |         tg128 |         92.34 ± 0.04 |
| llama 1B Q4_K - Medium         | 636.18 MiB |     1.10 B | Vulkan     |  99 |         tg128 |        110.08 ± 0.21 |
| llama 1B Q5_K - Medium         | 745.11 MiB |     1.10 B | Vulkan     |  99 |         tg128 |        110.75 ± 0.04 |
| llama 1B Q6_K                  | 860.86 MiB |     1.10 B | Vulkan     |  99 |         tg128 |        111.68 ± 0.27 |
| llama 1B Q8_0                  |   1.09 GiB |     1.10 B | Vulkan     |  99 |         tg128 |         96.21 ± 0.74 |