@netrunnereve 2024-11-18
On the other hand I'm able to get noticeably faster results on the 570 by adjusting the number of rows per workgroup. I wonder if it's worth making these things tweakable at some point like how we have K_QUANTS_PER_ITERATION.

| model                          |       size |     params | backend    | ngl | threads |          test |                  t/s | improvement from pre merge master |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------------: | -------------------: | ---: |
| llama 8B Q4_0 (1 row)                  |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          9.80 ± 0.05 | 27 % |
| llama 8B Q4_0 (2 row default)                  |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         11.60 ± 0.07 | 50% |
| llama 8B Q4_0 (8 rows)                 |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         12.64 ± 0.04 | 63% |
| llama 8B Q4_0 (16 rows)                 |   4.33 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |         12.49 ± 0.05 | 61% |

The numbers start going downhill once I go past 16 rows on Q4_0. Q8_0 also has some nice improvements as seen below.

| model                          |       size |     params | backend    | ngl | threads |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------: | ------------: | -------------------: |
| llama 8B Q8_0 (1 row default)                 |   7.95 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          6.60 ± 0.02 |
| llama 8B Q8_0 (2 rows)                 |   7.95 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          7.98 ± 0.00 |
| llama 8B Q8_0 (4 rows)                 |   7.95 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          9.38 ± 0.02 |
| llama 8B Q8_0 (8 rows)                 |   7.95 GiB |     8.03 B | Vulkan     | 100 |       8 |         tg128 |          9.27 ± 0.02 |